requirements.txt
google-adk==1.18.0
arxiv
requests
python-dotenv
google-generativeai
markdown
